Refactoring the common encryption and decryption code
