+++ app/authentication/jwt_decoder.py
@@ -1,16 +1,12 @@
 from cryptography.hazmat.backends.openssl.backend import backend
 from cryptography.hazmat.primitives import serialization
-from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
-from cryptography.hazmat.primitives import hashes
-from cryptography.hazmat.primitives.asymmetric import padding
 from cryptography.exceptions import InvalidTag, InternalError
-
+from app.cryptography.jwe_decryption import JWERSAOAEPDecryptor
 from app.authentication.invalid_token_exception import InvalidTokenException
 from app.authentication.no_token_exception import NoTokenException
 from app import settings
 from app.utilities import strings
 
-import base64
 import jwt
 import logging
 import json
@@ -19,7 +15,10 @@ IV_EXPECTED_LENGTH = 12
 CEK_EXPECT_LENGTH = 32
 
 
-class Decoder (object):
+class JWTDecryptor(JWERSAOAEPDecryptor):
+    '''
+    JWT signed with JWS RS256 And encrypted with JWE RSA-OAEP
+    '''
     def __init__(self):
         self.logger = logging.getLogger(__name__)
         if settings.EQ_USER_AUTHENTICATION_RRM_PUBLIC_KEY is None or settings.EQ_USER_AUTHENTICATION_SR_PRIVATE_KEY is None \
@@ -27,17 +26,11 @@ class Decoder (object):
                 self.logger.fatal('KEYMAT not configured correctly.')
                 raise OSError('KEYMAT not configured correctly.')
         else:
+            super().__init__(settings.EQ_USER_AUTHENTICATION_SR_PRIVATE_KEY)
             # oddly the python cryptography library needs these as bytes string
             rrm_public_key_as_bytes = strings.to_bytes(settings.EQ_USER_AUTHENTICATION_RRM_PUBLIC_KEY)
-            sr_private_key_as_bytes = strings.to_bytes(settings.EQ_USER_AUTHENTICATION_SR_PRIVATE_KEY)
-
             self.rrm_public_key = serialization.load_pem_public_key(
                 rrm_public_key_as_bytes,
-                backend=backend
-            )
-            self.sr_private_key = serialization.load_pem_private_key(
-                sr_private_key_as_bytes,
-                password=settings.EQ_USER_AUTHENTICATION_SR_PRIVATE_KEY_PASSWORD.encode(),
                 backend=backend
             )
 
@@ -155,8 +148,6 @@ class Decoder (object):
                 self.__check_jwe_protected_header(jwe_protected_header)
                 encrypted_key = tokens[1]
                 encoded_iv = tokens[2]
-                encoded_cipher_text = tokens[3]
-                encoded_tag = tokens[4]
 
                 decrypted_key = self._decrypt_key(encrypted_key)
                 iv = self._base64_decode(encoded_iv)
@@ -164,10 +155,8 @@ class Decoder (object):
                     raise InvalidTokenException("IV incorrect length")
                 if not self._check_cek_length(decrypted_key):
                     raise InvalidTokenException("CEK incorrect length")
-                tag = self._base64_decode(encoded_tag)
-                cipher_text = self._base64_decode(encoded_cipher_text)
 
-                signed_token = self._decrypt_cipher_text(cipher_text, iv, decrypted_key, tag, jwe_protected_header)
+                signed_token = super().decrypt(token)
                 return self.decode_signed_jwt_token(signed_token)
             else:
                 raise NoTokenException("JWT Missing")
@@ -191,25 +180,3 @@ class Decoder (object):
 
     def _check_cek_length(self, cek):
         return len(cek) == CEK_EXPECT_LENGTH
-
-    def _decrypt_key(self, encrypted_key):
-        decoded_key = self._base64_decode(encrypted_key)
-        key = self.sr_private_key.decrypt(decoded_key, padding.OAEP(mgf=padding.MGF1(algorithm=hashes.SHA1()), algorithm=hashes.SHA1(), label=None))
-        return key
-
-    def _decrypt_cipher_text(self, cipher_text, iv, key, tag, jwe_protected_header):
-        cipher = Cipher(algorithms.AES(key), modes.GCM(iv, tag), backend=backend)
-        decryptor = cipher.decryptor()
-        decryptor.authenticate_additional_data(jwe_protected_header.encode())
-        decrypted_token = decryptor.update(cipher_text) + decryptor.finalize()
-        return decrypted_token
-
-    @staticmethod
-    def _base64_decode(text):
-        # if the text is not a multiple of 4 pad with trailing =
-        # some base64 libraries don't pad data but Python is strict
-        # and will throw a incorrect padding error if we don't do this
-        if len(text) % 4 != 0:
-            while len(text) % 4 != 0:
-                text += "="
-        return base64.urlsafe_b64decode(text)
