+++ tests/app/authentication/test_jwt.py
@@ -4,7 +4,7 @@ import unittest
 
 from app import settings
 from app.authentication.invalid_token_exception import InvalidTokenException
-from app.authentication.jwt_decoder import Decoder
+from app.authentication.jwt_decoder import JWTDecryptor
 from tests.app.authentication import TEST_DO_NOT_USE_SR_PRIVATE_PEM
 
 # public key from jwt.io
@@ -34,13 +34,13 @@ class JWTTest(unittest.TestCase):
         settings.EQ_USER_AUTHENTICATION_SR_PRIVATE_KEY_PASSWORD = "digitaleq"
 
     def test_jwt_io(self):
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         token = decoder.decode_signed_jwt_token(jwtio_signed)
         self.assertEquals("John Doe", token.get("name"))
 
     def test_does_not_contain_two_instances_of_full_stop(self):
         jwe = jwtio_signed.replace('.', '', 1)
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwe)
         self.assertIn("Invalid Token", ite.exception.value)
@@ -48,7 +48,7 @@ class JWTTest(unittest.TestCase):
     def test_jwt_contains_empty_header(self):
         token_without_header = "e30." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(token_without_header)
         self.assertIn("Missing Headers", ite.exception.value)
@@ -56,21 +56,21 @@ class JWTTest(unittest.TestCase):
     def test_jwt_does_not_contain_header_at_all(self):
         token_without_header = "." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(token_without_header)
         self.assertIn("Missing Headers", ite.exception.value)
 
     def test_jwt_contains_empty_payload(self):
         token_without_payload = jwtio_header + ".e30." + jwtio_signature
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(token_without_payload)
         self.assertIn("Missing Payload", ite.exception.value)
 
     def test_jwt_does_not_contain_payload(self):
         token_without_payload = jwtio_header + ".." + jwtio_signature
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(token_without_payload)
         self.assertIn("Missing Payload", ite.exception.value)
@@ -78,7 +78,7 @@ class JWTTest(unittest.TestCase):
     def test_jwt_does_not_contain_signature(self):
         jwt = jwtio_header + "." + jwtio_payload + ".e30"
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Signature verification failed", ite.exception.value)
@@ -87,7 +87,7 @@ class JWTTest(unittest.TestCase):
         header = base64.urlsafe_b64encode(b'{"alg":"RS256", "kid":"EDCRRM"}')
         jwt = header.decode() + "." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Missing Type", ite.exception.value)
@@ -96,7 +96,7 @@ class JWTTest(unittest.TestCase):
         header = base64.urlsafe_b64encode(b'{"alg":"RS256", "kid":"EDCRRM", "typ":"TEST"}')
         jwt = header.decode() + "." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Invalid Type", ite.exception.value)
@@ -105,7 +105,7 @@ class JWTTest(unittest.TestCase):
         header = base64.urlsafe_b64encode(b'{"alg":"RS256", "kid":"EDCRRM","typ":"JWT","typ":"TEST"}')
         jwt = header.decode() + "." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Multiple typ Headers", ite.exception.value)
@@ -114,7 +114,7 @@ class JWTTest(unittest.TestCase):
         header = base64.urlsafe_b64encode(b'{"kid":"EDCRRM","typ":"JWT"}')
         jwt = header.decode() + "." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Missing Algorithm", ite.exception.value)
@@ -123,7 +123,7 @@ class JWTTest(unittest.TestCase):
         header = base64.urlsafe_b64encode(b'{"alg":"invalid","kid":"EDCRRM","typ":"JWT"}')
         jwt = header.decode() + "." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Invalid Algorithm", ite.exception.value)
@@ -132,7 +132,7 @@ class JWTTest(unittest.TestCase):
         header = base64.urlsafe_b64encode(b'{"alg":"None","kid":"EDCRRM","typ":"JWT"}')
         jwt = header.decode() + "." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Invalid Algorithm", ite.exception.value)
@@ -141,7 +141,7 @@ class JWTTest(unittest.TestCase):
         header = base64.urlsafe_b64encode(b'{"alg":"RS256", "alg":"HS256","kid":"EDCRRM", "typ":"JWT"}')
         jwt = header.decode() + "." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Multiple alg Headers", ite.exception.value)
@@ -150,7 +150,7 @@ class JWTTest(unittest.TestCase):
         header = base64.urlsafe_b64encode(b'{"alg":"RS256", "typ":"JWT"}')
         jwt = header.decode() + "." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Missing kid", ite.exception.value)
@@ -159,7 +159,7 @@ class JWTTest(unittest.TestCase):
         header = base64.urlsafe_b64encode(b'{"alg":"RS256", "kid":"EDCRRM", "kid":"test", "typ":"JWT"}')
         jwt = header.decode() + "." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Multiple kid Headers", ite.exception.value)
@@ -167,7 +167,7 @@ class JWTTest(unittest.TestCase):
     def test_signature_not_2048_bits(self):
         jwt = jwtio_header + "." + jwtio_payload + "." + base64.urlsafe_b64encode(os.urandom(255)).decode()
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Signature verification", ite.exception.value)
@@ -175,7 +175,7 @@ class JWTTest(unittest.TestCase):
     def test_payload_corrupt(self):
         jwt = jwtio_header + ".asdasd." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Corrupted Payload", ite.exception.value)
@@ -183,7 +183,7 @@ class JWTTest(unittest.TestCase):
     def test_header_corrupt(self):
         jwt = "asdsadsa" + "." + jwtio_payload + "." + jwtio_signature
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Corrupted Header", ite.exception.value)
@@ -191,7 +191,7 @@ class JWTTest(unittest.TestCase):
     def test_signature_corrupt(self):
         jwt = jwtio_header + "." + jwtio_payload + ".asdasddas"
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("DecodeError", ite.exception.value)
@@ -199,7 +199,7 @@ class JWTTest(unittest.TestCase):
     def test_payload_contains_malformed_json(self):
         payload = base64.urlsafe_b64encode(b'{"user":"jimmy,"iat": "1454935765","exp": "2075297148"')
         jwt = jwtio_header + "." + payload.decode() + "." + jwtio_signature
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("DecodeError", ite.exception.value)
@@ -207,7 +207,7 @@ class JWTTest(unittest.TestCase):
     def test_payload_contains_corrupted_json(self):
         payload = base64.urlsafe_b64encode(b'{"user":"jimmy","iat": "1454935765","exp": "2075297148"}ABDCE')
         jwt = jwtio_header + "." + payload.decode() + "." + jwtio_signature
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("DecodeError", ite.exception.value)
@@ -217,7 +217,7 @@ class JWTTest(unittest.TestCase):
                              "FtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWUsImlhdCI6IjE0NTQ5MzU3NjcifQ.VupTBEOEzeDjxd37PQ34xv" \
                              "BlLzeGTA0xFdGnLZDcnxAS1AjNcJ66edRmr4tmPIXnD6Mgen3HSB36xuXSnfzPld2msFHUXmB18CoaJQK19BXEY" \
                              "vosrBPzc1ohSvam_DgXCzdSMAcWSE63e6LTWNCT93-npD3p9tjdY_TWpEOOg14"
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(valid_token_no_exp)
         self.assertIn("Missing exp claim", ite.exception.value)
@@ -228,7 +228,7 @@ class JWTTest(unittest.TestCase):
                              "gdKQuNJ6o5oXA8rwEFiIE9UYo0yALhsQf5BbYs7RmVq760jde2FqwbwHze_XHFlOcg9nODfevEbRAUxjt_jpDaI" \
                              "LWmzreVw8jYTie9qn-F-7Tb6R1fgvvi5Fd7h0Py_LTLTZ72H1NOXCMtL_bbv6Y"
 
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(valid_token_no_iat)
         self.assertIn("Missing iat claim", ite.exception.value)
@@ -239,7 +239,7 @@ class JWTTest(unittest.TestCase):
                                        "HAiOiI_In0.0ApxEXw1rzo21XQo8WgcPvnz0e8QnT0GaoXVbCj-OdJtB7GArPzaiQ1cU53WaJsvGE" \
                                        "zHTczc6Y0xN7WzcTdcXN8Yjenf4VqoiYc6_FXGJ1s9Brd0JOFPyVipTFxPoWvYTWLXE-CAEpXrEb3" \
                                        "0kB3nRjHFV_yVhLiiZUU-gpUHqNQ"
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(valid_token_with_invalid_exp)
         self.assertIn("Expiration Time claim (exp) must be an integer", ite.exception.value)
@@ -250,7 +250,7 @@ class JWTTest(unittest.TestCase):
                                        "jk3MTQ4In0.1NIuxcD1FsZlU17NxK4UHdCfzl7qTV03qEaTRcqTC6A1Fs2Alc7mSQgkF_SpUw4Ylt" \
                                        "n-7DhO2InfcwDA0VhxBOHDL6ZzcEvzw-49iD-AaSd4aINIkDK-Iim5uzbKzgQCuZqSXFqxsZlezA4" \
                                        "BtwV7Lv2puqdPrXT8k3SvM2rOwRw"
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(valid_token_with_invalid_iat)
         self.assertIn("Issued At claim (iat) must be an integer", ite.exception.value)
@@ -261,7 +261,7 @@ class JWTTest(unittest.TestCase):
                                            "3NjUiLCJleHAiOiIxNDU0OTM1NzY2In0.PLIaSBh3jiPCsYgx7l1m8enorE7FzYUxVHgarlm" \
                                            "ZiMzpNjNmEzBYBq0yCk7wzkbrJhe5slliaMDY6C4hrAGo8oIUwYp_bQxxDCzyfeXiqdewdPe" \
                                            "L2X8D47Yw-KRt2XF03LXnMEyAaHD9CPhtnSWYUijka5h5yJIG62JTOGWvKGU"
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(valid_token_with_exp_in_the_past)
         self.assertIn("Signature has expired", ite.exception.value)
@@ -272,7 +272,7 @@ class JWTTest(unittest.TestCase):
                                              "5MzU3NjUiLCJleHAiOiIxNDU0OTM1NzY0In0.p2T1go1DblkgYiYHj_cLV1Jd169sG4bqv" \
                                              "qncTSpYPYawMIkjeb7s0IWFaswi348YLvcaAuQxaq3H5sw6RG3338TmFWYMJhvJTOYCZBC" \
                                              "pvkuu2PTKBAxYlQt9dHuWnFODYtjFgdsJigMrePdY9FIqyifyUAJsUHhA7WagOxUwW_I"
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(valid_token_with_exp_less_than_iat)
         self.assertIn("Signature has expired", ite.exception.value)
@@ -284,7 +284,7 @@ class JWTTest(unittest.TestCase):
                                          "iLCJleHAiOiIyMDc1Mjk3MTQ4In0.2MQNPXUAiT3AFhbQWoNZ3bs14WeQDloiJfP-3s9ddprzY" \
                                          "tj4omGaym32WvD-f4kBjuEwe479QzpJQoV_oTYCwB8VmFnKd4lKWKMUnNHQYcG0GWVB1Y9qWuW" \
                                          "nPG32kfS2Z7YuMkCgar8qttajB_YRcwhes4rIVaObFFXywnSinCQ"
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(valid_token_with_iat_in_future)
         self.assertIn("Issued At claim (iat) cannot be in the future", ite.exception.value)
@@ -295,7 +295,7 @@ class JWTTest(unittest.TestCase):
                                            b'"iat": "1454935765",'
                                            b'"exp": "2075297148"}')
         jwt = jwtio_header + "." + payload.decode() + "." + jwtio_signature
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Multiple iat claims", ite.exception.value)
@@ -306,7 +306,7 @@ class JWTTest(unittest.TestCase):
                                            b'"exp": "1454935765",'
                                            b'"exp": "2075297148"}')
         jwt = jwtio_header + "." + payload.decode() + "." + jwtio_signature
-        decoder = Decoder()
+        decoder = JWTDecryptor()
         with self.assertRaises(InvalidTokenException) as ite:
             decoder.decode_signed_jwt_token(jwt)
         self.assertIn("Multiple exp claims", ite.exception.value)
