+++ tests/app/authentication/test_jwt_decoder_using_rfc_example.py
@@ -1,7 +1,7 @@
 import unittest
 
 from app import settings
-from app.authentication.jwt_decoder import Decoder
+from app.authentication.jwt_decoder import JWTDecryptor
 
 # Not used in this test, but needed for the Decoder constructor
 private_pem = '''-----BEGIN RSA PRIVATE KEY-----
@@ -82,7 +82,7 @@ class JWTDecoderRFCTest(unittest.TestCase):
         self.assertEquals(plaintext, ''.join(chr(i) for i in plaintext_in_ascii))
 
     def test_decrypt(self):
-        decoder = Decoder()
+        decoder = JWTDecryptor()
 
         tokens = encrypted_jwt.split('.')
